server/requirements.txt
asgiref==3.4.1
click==8.0.1
ecdsa==0.17.0
fastapi==0.68.1
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
